pyproject.toml
[tool.poetry]
name = "flask-api"
version = "0.1.0"
description = ""
authors = ["USiqueira"]

[tool.poetry.dependencies]
python = "3.10.12"
gunicorn = "^21.2.0"
psycopg2-binary = "^2.9.9"
Flask = "^3.0.0"
Flask-SQLAlchemy = "^3.1.1"
Flask-Migrate = "^4.0.5"
Flask-RESTful = "^0.3.10"
python-decouple = "^3.8"
pytest = "^7.4.4"
Flask-JWT-Extended = "^4.6.0"
passlib = "^1.7.4"
flake8 = "^7.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
